EDU-332 refactor: 사전 평가 request 모델 분리
